requirements.txt

yt-dlp
faster-whisper
google-generativeai
python-dotenv
requests
Pillow
